+++ src/main/java/cn/nukkit/entity/item/EntityItem.java
@@ -183,7 +183,7 @@ public class EntityItem extends Entity {
         super.saveNBT();
         if (this.item != null) { // Yes, a item can be null... I don't know what causes this, but it can happen.
             this.namedTag.putCompound("Item", NBTIO.putItemHelper(this.item));
-            this.namedTag.putShort("Health", this.getHealth());
+            this.namedTag.putShort("Health", (int) this.getHealth());
             this.namedTag.putShort("Age", this.age);
             this.namedTag.putShort("PickupDelay", this.pickupDelay);
             if (this.owner != null) {
